Rush/requirements.txt
face_recognition==1.3.0
minio==7.2.0
numpy==1.21.6
opencv_python==4.6.0.66
pika==1.3.2
Pillow==9.2.0
Pillow==10.2.0
torch==1.8.1+cu101
torch==1.12.0
